Crate axsync

ArceOS synchronization primitives.

Currently supported primitives:

Mutex: A mutual exclusion primitive.

mod [spin]: spinlocks imported from the [kspin] crate.

Cargo Features

multitask: For use in the multi-threaded environments. If the feature is not enabled, Mutex will be an alias of [spin::SpinNoIrq]. This feature is enabled by default.

Re-exports

pub use kspin as spin;

Structs

Mutexmultitask

A mutual exclusion primitive useful for protecting shared data, similar to std::sync::Mutex.

MutexGuardmultitask

A guard that provides mutable data access.